requirements.txt
appdirs==1.4.4
backcall==0.2.0
black==20.8b1
blobfile==1.2.3
certifi==2021.5.30
-e git+https://github.com/openai/CLIP@8cad3a736a833bc4c9b4dd34ef12b52ec0e68856#egg=clip
cloudpickle==1.6.0
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
-e git+git@github.com:AranKomat/Diff-DALLE.git@649d8a7093b67a23befc79ee4bbdbcc43d373b1a#egg=diff_dalle
dill==0.3.3
filelock==3.0.12
ftfy==6.0.3
-e git+https://github.com/openai/guided-diffusion@912d5776a64a33e3baf3cff7eb1bcba9d9b9354c#egg=guided_diffusion
gym==0.18.3
imageio==2.9.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
kornia==0.5.6
matplotlib==3.4.2
matplotlib-inline==0.1.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpi4py==3.0.3
mypy-extensions==0.4.3
networkx==2.5.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271506491/work
olefile==0.46
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
pandarallel==1.5.2
pandas==1.2.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625655817137/work
prompt-toolkit==3.0.17
ptipython==1.0.1
ptpython==3.0.16
ptyprocess==0.7.0
pudb==2021.1
pycryptodomex==3.10.1
pyDeprecate==0.3.1
pyglet==1.5.15
Pygments==2.8.1
python-dateutil==2.8.1
python-magic==0.4.22
pytorch-lightning @ git+https://github.com/PytorchLightning/pytorch-lightning.git@a6fd32a70842058364861393c55891fd44e67a6e
pytorch-lightning-bolts==0.3.2
PyWavelets==1.1.1
pyzmq==22.1.0
regex==2021.7.6
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
taming-transformers-rom1504==0.0.6
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchvision==0.10.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.6
urwid==2.1.2
wcwidth==0.2.5
xmltodict==0.12.0
yapf==0.31.0
youtokentome==1.0.6
